+++ web/src/search/SearchAlert.tsx
@@ -6,6 +6,7 @@ interface Props {
     className: string
     title: string
     description?: string
+    errorBody?: React.ReactFragment
     proposedQueries?: GQL.ISearchQueryDescription[]
 
     location: H.Location
@@ -27,6 +28,7 @@ export class SearchAlert extends React.Component<Props, State> {
                     {<h2 className="search-alert__title">{this.props.title}</h2>}
                     {this.props.description && <p className="search-alert__description">{this.props.description}</p>}
                 </div>
+                {this.props.errorBody}
                 {this.props.proposedQueries && (
                     <ul className="search-alert__proposed-queries">
                         {this.props.proposedQueries.map((proposedQuery, i) => (